dev_requirements.txt
pylint>=2.5.3
pylint-django>=2.3.0
